refactor: reorganize tool configs
