+++ src/main.rs
@@ -2,6 +2,32 @@ use std::path::PathBuf;
 
 use clap::{Parser, Subcommand};
 
+/// Entrypoint.
+#[tokio::main]
+async fn main() -> std::io::Result<()> {
+    let cli = Cli::parse();
+
+    // Install tracing subscriber for logging.
+    let default_level = if cli.troubles { "debug" } else { "info" };
+    tracing_subscriber::fmt()
+        .with_env_filter(
+            tracing_subscriber::EnvFilter::try_from_default_env()
+                .unwrap_or_else(|_| tracing_subscriber::EnvFilter::new(default_level)),
+        )
+        .without_time()
+        .init();
+
+    match cli.command {
+        Commands::Init => aer::tool::init().await,
+        Commands::Palette => aer::tool::palette::run(),
+        Commands::Procs {
+            procs_file,
+            profile,
+        } => aer::tool::procs::run(procs_file.as_deref(), profile.as_deref()).await,
+    }
+}
+
+/// Top-level CLI arguments.
 #[derive(Parser)]
 #[command(name = "aer")]
 #[command(about = "A command-line toolkit for creatives", long_about = None)]
@@ -14,6 +40,7 @@ struct Cli {
     command: Commands,
 }
 
+/// CLI sub-commands.
 #[derive(Subcommand)]
 enum Commands {
     /// Initialize a new Aer project in the current directory
@@ -31,28 +58,4 @@ enum Commands {
         #[arg(short, long)]
         profile: Option<String>,
     },
-}
-
-#[tokio::main]
-async fn main() -> std::io::Result<()> {
-    let cli = Cli::parse();
-
-    // Install tracing subscriber for logging.
-    let default_level = if cli.troubles { "debug" } else { "info" };
-    tracing_subscriber::fmt()
-        .with_env_filter(
-            tracing_subscriber::EnvFilter::try_from_default_env()
-                .unwrap_or_else(|_| tracing_subscriber::EnvFilter::new(default_level)),
-        )
-        .without_time()
-        .init();
-
-    match cli.command {
-        Commands::Init => aer::tool::procs::init().await,
-        Commands::Palette => aer::tool::palette::run(),
-        Commands::Procs {
-            procs_file,
-            profile,
-        } => aer::tool::procs::run(procs_file.as_deref(), profile.as_deref()).await,
-    }
 }
